requirements.txt
nonebot2[fastapi,httpx,websockets]==2.3.1
nonebot-adapter-onebot==2.4.3
nonebot-plugin-apscheduler==0.4.0
loguru==0.7.2
httpx==0.27.0
chromadb==0.5.0
sentence-transformers==2.7.0
python-dotenv==1.0.1
torch>=2.0.0
numpy
pandas
fastapi
uvicorn
jinja2
python-multipart
# 根据需要添加
